+++ src/query/storages/common/cache-manager/src/caches.rs
@@ -24,7 +24,7 @@ use storages_common_cache::CacheAccessor;
 use storages_common_cache::InMemoryItemCacheHolder;
 use storages_common_cache::NamedCache;
 use storages_common_index::filters::Xor8Filter;
-use storages_common_index::BloomIndexMetaMini;
+use storages_common_index::BloomIndexMeta;
 use storages_common_table_meta::meta::SegmentInfo;
 use storages_common_table_meta::meta::TableSnapshot;
 use storages_common_table_meta::meta::TableSnapshotStatistics;
@@ -41,7 +41,7 @@ pub type TableSnapshotStatisticCache = NamedCache<InMemoryItemCacheHolder<TableS
 /// For each indexed data block, the bloom xor8 filter of column is cached individually
 pub type BloomIndexFilterCache = NamedCache<InMemoryItemCacheHolder<Xor8Filter>>;
 /// In memory object cache of parquet FileMetaData of bloom index data
-pub type BloomIndexMetaCache = NamedCache<InMemoryItemCacheHolder<BloomIndexMetaMini>>;
+pub type BloomIndexMetaCache = NamedCache<InMemoryItemCacheHolder<BloomIndexMeta>>;
 /// In memory object cache of parquet FileMetaData of external parquet files
 pub type FileMetaDataCache = NamedCache<InMemoryItemCacheHolder<FileMetaData>>;
 
@@ -87,7 +87,7 @@ impl CachedObject<TableSnapshotStatistics> for TableSnapshotStatistics {
     }
 }
 
-impl CachedObject<BloomIndexMetaMini> for BloomIndexMetaMini {
+impl CachedObject<BloomIndexMeta> for BloomIndexMeta {
     type Cache = BloomIndexMetaCache;
     fn cache() -> Option<Self::Cache> {
         CacheManager::instance().get_bloom_index_meta_cache()
