+++ src/query/storages/common/index/src/lib.rs
@@ -22,7 +22,7 @@ mod page_index;
 mod range_index;
 
 pub use bloom_index::BloomIndex;
-pub use bloom_index::BloomIndexMetaMini;
+pub use bloom_index::BloomIndexMeta;
 pub use bloom_index::FilterEvalResult;
 pub use index::Index;
 pub use page_index::PageIndex;
