+++ src/query/storages/fuse/src/io/read/bloom/column_filter_reader.rs
@@ -89,7 +89,7 @@ impl BloomColumnFilterReader {
             offset: meta.offset,
             len: meta.len,
             cache_key,
-            operator: operator.clone(),
+            operator,
             column_descriptor: ColumnDescriptor::new(
                 descriptor,
                 vec![column_name],
